Simplify user routes
