+++ routes/postRoutes.js
@@ -1,10 +1,10 @@
 const { Router } = require("express");
 
-const appPostController = require("../Controllers/postController");
+const postController = require("../controllers/postController");
 
 const router = Router();
 
-router.get("/", appPostController.getPosts);
-router.post("/", appPostController.addPost);
+router.get("/", postController.getPosts);
+router.post("/", postController.addPost);
 
 module.exports = router;
